+++ includes/orphaned-acf-media-refactored.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Orphaned ACF Media - Refactored Core Plugin Controller
  *